# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.Expense:
    properties:
      amount:
        description: cents
        type: integer
      category:
        description: |-
          Category is one of ExpenseCategorySlugs; it drives the expense's icon
          and grouping in the frontend.
        type: string
      created_at:
        type: string
      description:
        type: string
      group_id:
        type: integer
      id:
        type: integer
      items:
        items:
          $ref: '#/definitions/domain.ExpenseItem'
        type: array
      paid_by:
        allOf:
        - $ref: '#/definitions/domain.User'
        description: Relationships
      paid_by_id:
        type: integer
      receipt_image_url:
        type: string
      splits:
        items:
          $ref: '#/definitions/domain.ExpenseSplit'
        type: array
      updated_at:
        type: string
    type: object
  domain.ExpenseItem:
    properties:
      amount:
        description: cents
        type: integer
      description:
        type: string
      expense_id:
        type: integer
      id:
        type: integer
      users:
        description: Members this item is split among (equally).
        items:
          $ref: '#/definitions/domain.User'
        type: array
    type: object
  domain.ExpenseSplit:
    properties:
      amount:
        description: cents
        type: integer
      created_at:
        type: string
      expense_id:
        type: integer
      id:
        type: integer
      updated_at:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/domain.User'
        description: Relationships
      user_id:
        type: integer
    type: object
...
  domain.User:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      push_enabled:
        description: |-
          PushEnabled is the user's own preference, checked before sending any
          push notification — separate from whether they have any subscribed
          devices (PushSubscription rows).
        type: boolean
      updated_at:
        type: string
    type: object
...
  handler.AddExpenseRequest:
    properties:
      amount:
        example: 12050
        type: integer
      category:
        example: food
        type: string
      description:
        example: Dinner
        type: string
      group_id:
        example: 1
        type: integer
      items:
        items:
          $ref: '#/definitions/handler.ItemInputRequest'
        type: array
      paid_by_id:
        example: 1
        type: integer
      receipt_image_path:
        description: |-
          ReceiptImagePath, if set, is the storage path returned by a prior
          POST /receipts/scan call — carries the scanned image over to the
          expense it was scanned for.
        type: string
      split_method:
        enum:
        - equal
        - percentage
        - fixed
        - shares
        example: equal
        type: string
      splits:
        items:
          $ref: '#/definitions/handler.SplitInputRequest'
        type: array
    type: object
...
  handler.ItemInputRequest:
    properties:
      amount:
        example: 1500
        type: integer
      description:
        example: Burger
        type: string
      user_ids:
        items:
          type: integer
        type: array
    type: object
...
  handler.SplitInputRequest:
    properties:
      user_id:
        example: 2
        type: integer
      value:
        example: 50
        type: number
    type: object
...
  handler.UpdateExpenseRequest:
    properties:
      amount:
        example: 12050
        type: integer
      category:
        example: food
        type: string
      description:
        example: Dinner
        type: string
      items:
        items:
          $ref: '#/definitions/handler.ItemInputRequest'
        type: array
      paid_by_id:
        example: 1
        type: integer
      receipt_image_path:
        description: |-
          ReceiptImagePath: omit the field to leave the expense's existing image
          (if any) untouched; include it (even as "") to replace it.
        type: string
      split_method:
        enum:
        - equal
        - percentage
        - fixed
        - shares
        example: equal
        type: string
      splits:
        items:
          $ref: '#/definitions/handler.SplitInputRequest'
        type: array
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API server for SplitEasy app (Splitwise style expense sharing).
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: SplitEasy API
  version: "1.0"
paths:
...
  /expenses:
    post:
      consumes:
      - application/json
      description: Adds a new expense to a group and splits it. split_method can be
        "equal" (default; splits among all members, or the given subset), "percentage"
        (splits values must add up to 100), "fixed" (splits values must add up to
        amount), or "shares" (splits proportionally to relative weights).
      parameters:
      - description: Expense Info
        in: body
        name: expense
        required: true
        schema:
          $ref: '#/definitions/handler.AddExpenseRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Expense'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - JWT: []
      summary: Add an expense
      tags:
      - expenses
  /expenses/{id}:
    delete:
      description: Soft-deletes an expense. Only the payer or a split participant
        may delete it.
      parameters:
      - description: Expense ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - JWT: []
      summary: Delete an expense
      tags:
      - expenses
    get:
      description: Retrieves a single expense by id. Any member of the expense's group
        may view it — unlike edit/delete, this isn't limited to the payer or split
        participants.
      parameters:
      - description: Expense ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Expense'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - JWT: []
      summary: Get an expense
      tags:
      - expenses
    put:
      consumes:
      - application/json
      description: Replaces an expense's payer, description, amount, split, and items.
        Only the current payer or a current split participant may edit it.
      parameters:
      - description: Expense ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated expense info
        in: body
        name: expense
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateExpenseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Expense'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      security:
      - JWT: []
      summary: Edit an expense
      tags:
      - expenses
...
  /groups/{groupId}/expenses:
    get:
      description: Retrieves list of expenses for a specific group.
      parameters:
      - description: Group ID
        in: path
        name: groupId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Expense'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - JWT: []
      summary: Get expenses of a group
      tags:
      - expenses
...
securityDefinitions:
  JWT:
    description: Type "Bearer <your-jwt-token>" to authenticate.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"